Separate level handling from Game
For some effects, like animations when the player dies or wins, we need
a good separation from the game code.

For this the pattern of a finite state machine is suitable. To realize
this, we move in this commit the functionality from Game to
GameStateLevel (-> The player actually plays in this state). The Games
object now takes over the management of the states.