Introduce tick_time
During calculation of a single tick the objects were individually
calling `get_ms()` or `level->time()` to get the current time and
perform some calculation.

This approach was working, but made the synchronization of different
objects problematic. However, with the new level 6 we want some
enemies to move completely synchronous.

To make this possible, we now once call `get_ms()` for each tick and
propagate the result down the call stack as `tick_time`.